requirements.txt
asgiref==3.8.1
Django==4.2.16
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
